+++ lib/zenlish/wclasses/lexical_verb.rb
@@ -2,8 +2,10 @@ require_relative 'verb'
 
 module Zenlish
   module WClasses
-    # TODO: document
     class LexicalVerb < Verb
+      def initialize
+        super
+      end
     end # class
   end # module
 end # module