+++ spec/zenlish/inflect/inflection_table_builder_spec.rb
@@ -97,7 +97,7 @@ module Zenlish
 
         MockFeatureBearer = Struct.new(:NUMBER, :base_form)
 
-        it 'should build tables conform to specs' do
+        it 'should build default paradigm for common nouns' do
           table = subject.build('Common_form') do
             feature_heading 'NUMBER'
             method_heading 'base_form'
@@ -121,6 +121,86 @@ module Zenlish
           mck_3['NUMBER'] = :plural
           expect(table.inflect(mck_3, [nil, nil])).to eq('bodies')
         end
+        
+        MockRegularVerb = Struct.new(:base_form, :PERSON, :NUMBER, :TIME)
+        it 'should build default paradigm for regular verbs' do
+          table = subject.build('Regular_form') do
+            feature_heading 'PERSON'
+            feature_heading 'NUMBER'
+            feature_heading 'TIME'
+            method_heading  'base_form'
+            #     PERSON             NUMBER             TIME                         base_form
+            rule([not_equal(:third), dont_care,         equals(:present),            dont_care], col('base_form'))
+            rule([equals(:third),    equals(:singular), equals(:present),            matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ies'))
+            rule([equals(:third),    equals(:singular), equals(:present),            matches(/(?:[osxz]|ch|sh)$/)], concat(col('base_form'), 'es'))             
+            rule([equals(:third),    equals(:singular), equals(:present),            dont_care], concat(col('base_form'), 's'))         
+            rule([dont_care,         dont_care, equals(:progressive),                matches(/e$/)], sub(col('base_form'), /e$/, 'ing'))            
+            rule([dont_care,         dont_care, equals(:progressive),                dont_care], concat(col('base_form'), 'ing'))
+            rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/e$/)], concat(col('base_form'), 'd'))             
+            rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ied'))            
+            rule([dont_care,         dont_care, in?(:past_simple, :past_participle), dont_care], concat(col('base_form'), 'ed'))
+          end
+          
+          verb_1 = MockRegularVerb.new('belong', :first, :singular, :present)
+          expect(table.inflect(verb_1, [nil,    nil,     nil, nil]             )).to eq('belong')
+          expect(table.inflect(verb_1, [:third, nil,     nil, nil]             )).to eq('belongs')
+          expect(table.inflect(verb_1, [nil,    :plural, nil, nil]             )).to eq('belong')
+          expect(table.inflect(verb_1, [nil,    nil,     :progressive, nil]    )).to eq('belonging')
+          expect(table.inflect(verb_1, [nil,    nil,     :past_simple, nil]    )).to eq('belonged')
+          expect(table.inflect(verb_1, [nil,    nil,     :past_participle, nil])).to eq('belonged')
+          
+          verb_2 = MockRegularVerb.new('try', :first, :singular, :present)
+          expect(table.inflect(verb_2, [nil,    nil,     nil, nil]             )).to eq('try')
+          expect(table.inflect(verb_2, [:third, nil,     nil, nil]             )).to eq('tries')
+          expect(table.inflect(verb_2, [nil,    :plural, nil, nil]             )).to eq('try')
+          expect(table.inflect(verb_2, [nil,    nil,     :progressive, nil]    )).to eq('trying')
+          expect(table.inflect(verb_2, [nil,    nil,     :past_simple, nil]    )).to eq('tried')
+          expect(table.inflect(verb_2, [nil,    nil,     :past_participle, nil])).to eq('tried')
+
+          verb_3 = MockRegularVerb.new('touch', :first, :singular, :present)
+          expect(table.inflect(verb_3, [:second,nil,     nil, nil]             )).to eq('touch')
+          expect(table.inflect(verb_3, [:third, nil,     nil, nil]             )).to eq('touches')
+          expect(table.inflect(verb_3, [nil,    :plural, nil, nil]             )).to eq('touch')
+          expect(table.inflect(verb_3, [nil,    nil,     :progressive, nil]    )).to eq('touching')
+          expect(table.inflect(verb_3, [nil,    nil,     :past_simple, nil]    )).to eq('touched')
+          expect(table.inflect(verb_3, [nil,    nil,     :past_participle, nil])).to eq('touched')
+
+          verb_4 = MockRegularVerb.new('cause', :first, :singular, :present)
+          expect(table.inflect(verb_4, [nil,    nil,     nil, nil]             )).to eq('cause')
+          expect(table.inflect(verb_4, [:third, nil,     nil, nil]             )).to eq('causes')
+          expect(table.inflect(verb_4, [nil,    :plural, nil, nil]             )).to eq('cause')
+          expect(table.inflect(verb_4, [nil,    nil,     :progressive, nil]    )).to eq('causing')
+          expect(table.inflect(verb_4, [nil,    nil,     :past_simple, nil]    )).to eq('caused')
+          expect(table.inflect(verb_4, [nil,    nil,     :past_participle, nil])).to eq('caused')          
+        end
+        
+        # MockIrregularVerb = Struct.new(:base_form, :PERSON, :NUMBER, :TIME)
+        # it 'should build default paradigm for irregular verbs' do
+          # table = subject.build('Irregular_form') do
+            # feature_heading 'PERSON'
+            # feature_heading 'NUMBER'
+            # feature_heading 'TIME'
+            # method_heading  'base_form'
+            # #     PERSON             NUMBER             TIME                         base_form
+            # rule([not_equal(:third), dont_care,         equals(:present),            dont_care], col('base_form'))
+            # rule([equals(:third),    equals(:singular), equals(:present),            matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ies'))
+            # rule([equals(:third),    equals(:singular), equals(:present),            matches(/(?:[osxz]|ch|sh)$/)], concat(col('base_form'), 'es'))             
+            # rule([equals(:third),    equals(:singular), equals(:present),            dont_care], concat(col('base_form'), 's'))         
+            # rule([dont_care,         dont_care, equals(:progressive),                matches(/e$/)], sub(col('base_form'), /e$/, 'ing'))            
+            # rule([dont_care,         dont_care, equals(:progressive),                dont_care], concat(col('base_form'), 'ing'))
+            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/e$/)], concat(col('base_form'), 'd'))             
+            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ied'))            
+            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), dont_care], concat(col('base_form'), 'ed'))
+          # end
+
+          # verb_1 = MockIrregularVerb.new('choose', :first, :singular, :present)
+          # expect(table.inflect(verb_1, [nil,    nil,     nil, nil]             )).to eq('choose')
+          # expect(table.inflect(verb_1, [:third, nil,     nil, nil]             )).to eq('chooses')
+          # expect(table.inflect(verb_1, [nil,    :plural, nil, nil]             )).to eq('choose')
+          # expect(table.inflect(verb_1, [nil,    nil,     :progressive, nil]    )).to eq('choosing')
+          # expect(table.inflect(verb_1, [nil,    nil,     :past_simple, nil]    )).to eq('chose')
+          # expect(table.inflect(verb_1, [nil,    nil,     :past_participle, nil])).to eq('chosen')
+        # end
       end # context
     end # describe
   end # module